Lab 4 Started

+++ Labs/Lab4/Account.java
@@ -40,7 +40,7 @@ class Account
 
 	/* Returns the balance of the Account
 	 */
-	public int getBalance()
+	public double getBalance()
 	{
 		return balance;
 	}
